Move cases class to its own file
